more complete overhaul for new CircleCI testing regime
